requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.14.0
django-debug-toolbar==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djangorestframework-xml==2.0.0
djoser==2.1.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13

# asgiref==3.5.2
# certifi==2022.9.24
# cffi==1.15.1
# chardet==3.0.4
# charset-normalizer==2.1.1
# coreapi==2.3.3
# coreschema==0.0.4
# cryptography==38.0.4
# # cx-Oracle==8.3.0
# defusedxml==0.7.1
# Django==4.1.3
# django-debug-toolbar==3.7.0
# django-templated-mail==1.1.1
# djangorestframework==3.14.0
# djangorestframework-simplejwt==4.8.0
# djangorestframework-xml==2.0.0
# djoser==2.1.0
# gunicorn==20.1.0
# h11==0.9.0
# h2==3.2.0
# hpack==3.0.0
# hstspreload==2023.1.1
# httpcore==0.9.1
# httpx==0.13.3
# hyperframe==5.2.0
# idna==2.10
# itypes==1.2.0
# Jinja2==3.1.2
# MarkupSafe==2.1.1
# oauthlib==3.2.2
# psycopg2-binary
# pycparser==2.21
# PyJWT==2.6.0
# python3-openid==3.2.0
# pytz==2022.6
# requests==2.28.1
# requests-oauthlib==1.3.1
# rfc3986==1.5.0
# six==1.16.0
# sniffio==1.3.0
# social-auth-app-django==4.0.0
# social-auth-core==4.3.0
# sqlparse==0.4.3
# uritemplate==4.1.1
# urllib3==1.26.13
